+++ src/TetrisBoard.py
@@ -1,20 +1,5 @@
 from src.PygameDelegate import PygameDelegate
 pygame = PygameDelegate()
-from src.RandomDelegate import RandomDelegate
-random = RandomDelegate()
-
-NUM_OF_SHAPE_GRID_ROWS = 4
-NUM_OF_SHAPE_GRID_COLUMNS = 4
-ACCOUNT_FOR_NEXT_ROW = 4
-
-COLORS = (
-    (120, 37, 179),
-    (100, 179, 179),
-    (80, 34, 22),
-    (80, 134, 22),
-    (180, 34, 22),
-    (180, 34, 122),
-)
 
 BLACK = (0, 0, 0)
 WHITE = (255, 255, 255)
@@ -23,159 +8,23 @@ GRAY = (128, 128, 128)
 # codesmell - params should be assinged to a var and var used instead of param
 # codesmell - long class
 class TetrisBoard:
-    def __init__(self, tetris_block):
-        self.tetris_block = tetris_block
+    def __init__(self):
         self.size_of_grid_block = 20
         self.start_x_position = 100
         self.start_y_position = 60
         self.grid_block_height = 0
         self.grid_block_width = 0        
-        self.current_figure_type = 0
-        self.current_figure_color = 0
-        self.current_rotation = 0
-        self.shift_in_x = 0
-        self.shift_in_y = 0
         self.window_size = (400, 500)
-        self.game_state = "start"  # or "gameover"
         self.game_field = []
-        self.Figures = (
-            [(1, 5, 9, 13), (4, 5, 6, 7)],
-            [(4, 5, 9, 10), (2, 6, 5, 9)],
-            [(6, 7, 9, 10), (1, 5, 6, 10)],
-            [(1, 2, 5, 9), (0, 4, 5, 6), (1, 5, 9, 8), (4, 5, 6, 10)],
-            [(1, 2, 6, 10), (5, 6, 7, 9), (2, 6, 10, 11), (3, 5, 6, 7)],
-            [(1, 4, 5, 6), (1, 4, 5, 9), (4, 5, 6, 9), (1, 5, 6, 9)],
-            [(1, 2, 5, 6)],
+        self.colors = (
+            (120, 37, 179),
+            (100, 179, 179),
+            (80, 34, 22),
+            (80, 134, 22),
+            (180, 34, 22),
+            (180, 34, 122),
         )
 
-    def create_figure(self, starting_shift_x, starting_shift_y):
-        self.set_shift_in_x(starting_shift_x)
-        self.set_shift_in_y(starting_shift_y)
-        self.tetris_block.create_figure()
-        if self.check_intersection():
-            self.set_game_state("gameover")
-
-
-
-    def check_intersection(self):
-        is_intersection = False
-        game_field = self.get_game_field()
-        shift_in_x = self.get_shift_in_x()
-        shift_in_y = self.get_shift_in_y()
-        grid_block_height = self.get_grid_block_height()
-        grid_block_width = self.get_grid_block_width()
-        shape = self.get_shape()
-
-        for current_row in range (NUM_OF_SHAPE_GRID_ROWS) :
-            for current_column in range (NUM_OF_SHAPE_GRID_COLUMNS) :
-                if current_row * ACCOUNT_FOR_NEXT_ROW  + current_column in shape:
-                    is_emplty = 0
-                    if current_row + shift_in_y > grid_block_height - 1 or \
-                    current_column + shift_in_x > grid_block_width - 1 or \
-                    current_column + shift_in_x < is_emplty or \
-                    game_field[current_row + shift_in_y][ current_column  + shift_in_x] > is_emplty:
-                        is_intersection = True
-        return is_intersection
-   
-    """The freeze_figure method has a bug where it looks like it is not freezing a figure to the board.
-        This bug does not appear to be consistent, with most figures being frozen as they should.
-        This bug involves the create_figure method. This bug happens when you change create_figure from
-        a global method to either an internal method or a called method. I have not been able to figure out why,
-        but all other functions work. I think it may have to do with the refreshing of the game."""
-    # maybe make a board_managment class with freeze, clear, check if filled, delete and all moves (maybe draw_figure)
-    def freeze_figure(self):
-        game_field = self.get_game_field()
-        shape = self.get_shape()
-        shift_in_x = self.get_shift_in_x()
-        shift_in_y = self.get_shift_in_y()
-        shape_color = self.tetris_block.get_current_figure_color()
-
-        for current_row in range (NUM_OF_SHAPE_GRID_ROWS) :
-            for current_column in range (NUM_OF_SHAPE_GRID_COLUMNS) :
-                if current_row * ACCOUNT_FOR_NEXT_ROW  + current_column in shape:
-                    game_field[ current_row + shift_in_y][ current_column  + shift_in_x] = shape_color
-
-    def clear_lines(self):
-        grid_block_height = self.get_grid_block_height()
-        start_of_range = 1
-    
-        for current_row in range(start_of_range, grid_block_height):
-            if self.check_if_row_is_filled(current_row):
-                self.delete_row(current_row)
-
-    def check_if_row_is_filled(self, current_row):
-        grid_block_width = self.get_grid_block_width()
-        game_field = self.get_game_field()
-        is_empty = 0
-        for current_column in range(grid_block_width):
-            if game_field[ current_row ][ current_column ] == is_empty:
-                is_empty += 1
-        return is_empty == 0
-            
-    def delete_row(self, current_row):
-        game_field = self.get_game_field()
-        grid_block_width = self.get_grid_block_width()
-        end_of_range = 1
-        size_of_step_through_range = -1
-        for current_row_above in range(current_row, end_of_range, size_of_step_through_range):
-            for current_column in range(grid_block_width):
-                game_field[current_row_above][ current_column ] = game_field[current_row_above - 1][ current_column ]
-    # maybe make moves sub classes of a moves class and use poly
-    
-    def game_rules(self):
-        starting_shift_x = 3
-        starting_shift_y = 0  
-        self.freeze_figure()
-        self.create_figure(starting_shift_x, starting_shift_y)
-        self.clear_lines() 
-    
-    def move_to_bottom(self):
-        
-        while not self.check_intersection():
-            self.set_shift_in_y(self.get_shift_in_y() + 1)
-        self.set_shift_in_y(self.get_shift_in_y() - 1)
-        self.game_rules()
-
-    def move_down(self):
-        self.set_shift_in_y(self.get_shift_in_y() + 1)
-        if self.check_intersection():
-            self.set_shift_in_y(self.get_shift_in_y() - 1)
-            self.game_rules()
-
-
-    def move_sideways(self, player_change_in_x):
-        shift_in_x = self.get_shift_in_x()
-
-        self.set_shift_in_x(shift_in_x + player_change_in_x)
-        if self.check_intersection():
-            self.set_shift_in_x(shift_in_x)
-
-    def rotate_figure(self):      
-        current_rotation = self.tetris_block.get_current_rotation()
-        figure = self.tetris_block.get_figure() # add geter to tetris block for this.
-        current_figure_type = self.tetris_block.get_current_figure_type()
-        
-        self.tetris_block.set_current_rotation((self.tetris_block.get_current_rotation() + 1) % len(figure[current_figure_type]))
-        if self.check_intersection():
-            self.tetris_block.set_current_rotation((self.tetris_block.get_current_rotation() - 1)  % len(figure[current_figure_type]))
-
-    def draw_figure(self, screen):
-        shape = self.get_shape()
-        color = COLORS[self.tetris_block.get_current_figure_color()]
-        start_x_position = self.get_start_x_position()
-        start_y_position = self.get_start_y_position()
-        size_of_grid_block = self.get_size_of_grid_block()
-        shift_in_x = self.get_shift_in_x()
-        shift_in_y = self.get_shift_in_y()
-
-        for current_row in range (NUM_OF_SHAPE_GRID_ROWS) :
-            for current_column in range (NUM_OF_SHAPE_GRID_COLUMNS) :
-                position = current_row * ACCOUNT_FOR_NEXT_ROW  +  current_column 
-                if position in shape:
-                    pygame.draw.rect(screen, color,
-                                    [start_x_position  + size_of_grid_block * ( current_column  + shift_in_x) + 1,
-                                    start_y_position + size_of_grid_block * ( current_row + shift_in_y) + 1,
-                                    size_of_grid_block - 2, size_of_grid_block - 2])
 
     # make a class that connects pygame to this one and remove pygame form class
     def draw_game_board(self, screen):
@@ -192,7 +41,7 @@ class TetrisBoard:
             for current_column in range(grid_block_width):
                 pygame.draw.rect(screen, GRAY, [start_x_position + size_of_grid_block *  current_column , start_y_position + size_of_grid_block * current_row, size_of_grid_block, size_of_grid_block], 1)
                 if game_field[ current_row ][ current_column ] > 0:
-                    pygame.draw.rect(screen, COLORS[game_field[ current_row ][ current_column ]],
+                    pygame.draw.rect(screen, self.colors[game_field[ current_row ][ current_column ]],
                                     [start_x_position + size_of_grid_block * current_column + 1, start_y_position + size_of_grid_block * current_row + 1, size_of_grid_block - 2, size_of_grid_block - 1])
 
     def initialize_board(self, height, width):
@@ -200,7 +49,6 @@ class TetrisBoard:
         self.set_grid_block_height(height)
         self.set_grid_block_width(width)
         self.set_game_field([])
-        self.set_game_state("start")
         grid_block_height = self.get_grid_block_height()
         grid_block_width = self.get_grid_block_width()
         is_empty = 0
@@ -243,61 +91,13 @@ class TetrisBoard:
 
     def set_grid_block_width(self, value):
         self.grid_block_width = value
-    
-        # Getter and setter methods for current_figure_type
-    def get_current_figure_type(self):
-        return self.current_figure_type
 
-    def set_current_figure_type(self, value):
-        self.current_figure_type = value
-
-    # Getter and setter methods for current_figure_color
-    def get_current_figure_color(self):
-        return self.current_figure_color
-
-    def set_current_figure_color(self, value):
-        self.current_figure_color = value
-
-    # Getter and setter methods for current_rotation
-    def get_current_rotation(self):
-        return self.current_rotation
-
-    def set_current_rotation(self, value):
-        self.current_rotation = value
-
-    # Getter and setter methods for shift_in_x
-    def get_shift_in_x(self):
-        return self.shift_in_x
-
-    def set_shift_in_x(self, value):
-        self.shift_in_x = value
-
-    # Getter and setter methods for shift_in_y
-    def get_shift_in_y(self):
-        return self.shift_in_y
-
-    def set_shift_in_y(self, value):
-        self.shift_in_y = value
-
-    def get_shape(self):
-        return self.tetris_block.get_shape()
-    
         # Getter and setter for window_size
     def get_window_size(self):
         return self.window_size
 
     def set_window_size(self, value):
         self.window_size = value
-
-    # Getter and setter for game_state
-    def get_game_state(self):
-        return self.game_state
-
-    def set_game_state(self, value):
-        if value == "start" or value == "gameover":
-            self.game_state = value
-        else:
-            self.game_state = "gameover"
 
     # Getter and setter for game_field
     def get_game_field(self):
@@ -309,3 +109,14 @@ class TetrisBoard:
     # Method to append to game_field
     def append_to_game_field(self, value):
         self.game_field.append(value)
+
+        # Getter and setter for colors
+    def get_colors(self):
+        return self.colors
+
+    def set_colors(self, value):
+        self.colors = value
+
+    # Method to append to colors
+    def append_to_colors(self, value):
+        self.colors.append(value)
