+++ tetris_verFinal.py
@@ -1,17 +1,21 @@
 from src.PygameDelegate import PygameDelegate
 from src.TetrisBoard import TetrisBoard
 from src.TetrisBlock import TetrisBlock
+from src.TetrisBoardManager import BoardManager
+from src.TetrisBoardChecker import BoardChecker
 pygame = PygameDelegate()
 
 
 def main():   
 
-    block = TetrisBlock()
-    board = TetrisBoard(block)
+    tetris_block = TetrisBlock()
+    tetris_board = TetrisBoard()
+    board_manager = BoardManager(tetris_block, tetris_board)
+    board_checker = BoardChecker(tetris_board)
  
     # Pygame related init
     pygame.init()
-    screen = pygame.display.set_mode(board.get_window_size())
+    screen = pygame.display.set_mode(board_manager.get_window_size())
     pygame.display.set_caption("Tetris")
     clock = pygame.time.Clock()
 
@@ -23,25 +27,24 @@ def main():
     game_block_height = 20
     game_block_width = 10
 
-    board.initialize_board(game_block_height, game_block_width) # code smell - what is 20 and 10? Can we use keyword argument? 
+    tetris_board.initialize_board(game_block_height, game_block_width) # code smell - what is 20 and 10? Can we use keyword argument? 
     
     starting_shift_x = 3 
     starting_shift_y = 0
 
-    board.create_figure(starting_shift_x, starting_shift_y)
+    board_manager.create_figure(starting_shift_x, starting_shift_y)
     done = False
     level = 1
     interval = 100000
     interval_of_auto_move = 2
 
     event_key_action_list = {
-        pygame.K_UP: lambda: board.rotate_figure(),
+        pygame.K_UP: lambda: board_manager.rotate_figure(),
         pygame.K_DOWN: "true",
-        pygame.K_LEFT: lambda: board.move_sideways(-1),
-        pygame.K_RIGHT: lambda: board.move_sideways(1),
-        pygame.K_SPACE: lambda: board.move_to_bottom()
+        pygame.K_LEFT: lambda: board_manager.move_sideways(-1),
+        pygame.K_RIGHT: lambda: board_manager.move_sideways(1),
+        pygame.K_SPACE: lambda: board_manager.move_to_bottom()
     }
-
 
     while not done:
         counter += 1
@@ -50,8 +53,8 @@ def main():
             
         # Check if we need to automatically go down
         if counter % (fps // interval_of_auto_move // level) == 0 or pressing_down: 
-            if board.get_game_state() == "start":
-                board.move_down()
+            if board_manager.get_game_state() == "start":
+                board_manager.move_down()
 
         for event in pygame.event.get():
             if event.type == pygame.QUIT:
@@ -63,17 +66,17 @@ def main():
                     method_to_run = event_key_action_list[event.key]
                     if callable(method_to_run):
                         method_to_run()
-
+                        
 
             if event.type == pygame.KEYUP and event.key == pygame.K_DOWN:
                 pressing_down = False
                 
-        board.draw_game_board(screen = screen)
+        tetris_board.draw_game_board(screen = screen)
         
         # code smell - how many values duplication Figures[current_figure_type][current_rotation]
-        board.draw_figure(screen = screen)
-
-        if board.get_game_state() == "gameover":
+        board_manager.draw_figure(screen = screen)
+        board_checker.clear_lines()
+        if board_manager.get_game_state() == "gameover":
             done = True
 
         # refresh the screen
